Move reusable tab code to new component file
